🐛 ️ ️[#122] 찜하기 반환 메시지 전달 방식 변경 (#128)
* 🐛 ️ ️[#122] 122번만 다시 커밋

* [#122] 122번 LikeStatus부분 수정

* [#122] LikeServiceImpl 클래스 삭제

+++ goodsending/src/main/java/com/goodsending/productlike/service/LikeService.java
@@ -11,6 +11,7 @@ import com.goodsending.productlike.dto.LikeRequestDto;
 import com.goodsending.productlike.dto.LikeResponseDto;
 import com.goodsending.productlike.entity.Like;
 import com.goodsending.productlike.repository.LikeRepository;
+import com.goodsending.productlike.type.LikeStatus;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -46,16 +47,36 @@ public class LikeService {
         like = new Like(product, member);
         likeRepository.save(like);
         countLike(product);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
+        LikeResponseDto likeResponseDto = LikeResponseDto.of(HttpStatus.CREATED,
+            LikeStatus.CREATE_SUCCESS);
+        return ResponseEntity.ok(likeResponseDto);
+        // 버튼이 true이고 찜이 이미 존재하면 찜하기 실패, BAD_REQUEST, ALEADYLIKE 발생
+      } else {
+        LikeResponseDto likeResponseDto = LikeResponseDto.of(HttpStatus.BAD_REQUEST,
+            LikeStatus.ALREADY_LIKE);
+        return ResponseEntity.ok(likeResponseDto);
       }
     } else {
-      like = likeRepository.findLikeByMemberAndProduct(member,
-          product).orElseThrow(() -> CustomException.from(ExceptionCode.MEMBER_NOT_FOUND));
-      likeRepository.delete(like);
-      countLike(product);
-      return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+      // 찜이 존재하지 않으면 삭제 실패, BAD_REQUEST, FAIL 발생
+      if (!existingLike) {
+        LikeResponseDto likeResponseDto = LikeResponseDto.of(HttpStatus.BAD_REQUEST,
+            LikeStatus.DELETED_LIKE);
+        return ResponseEntity.ok(likeResponseDto);
+        // 찜이 존재하면 삭제 성공, NO_CONTENT, SUCCESS 발생
+      } else {
+        like = findLikeByMemberAndProduct(member, product);
+        likeRepository.delete(like);
+        countLike(product);
+        LikeResponseDto likeResponseDto = LikeResponseDto.of(HttpStatus.NO_CONTENT,
+            LikeStatus.REMOVE_SUCCESS);
+        return ResponseEntity.ok(likeResponseDto);
+      }
     }
-    return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+  }
+
+  private Like findLikeByMemberAndProduct(Member member, Product product) {
+    return likeRepository.findLikeByMemberAndProduct(member,
+        product).orElseThrow(() -> CustomException.from(ExceptionCode.MEMBER_NOT_FOUND));
   }
 
   private void countLike(Product product) {
